+++ lib/knapsack_pro/build_distribution_fetcher.rb
@@ -43,7 +43,8 @@ module KnapsackPro
         commit_hash: repository_adapter.commit_hash,
         branch: repository_adapter.branch,
         node_total: KnapsackPro::Config::Env.ci_node_total,
-        node_index: KnapsackPro::Config::Env.ci_node_index
+        node_index: KnapsackPro::Config::Env.ci_node_index,
+        test_queue_id: KnapsackPro::Config::Env.test_queue_id_
       }.merge(additional_params)
 
       KnapsackPro::Client::API::V1::BuildDistributions.last(request_hash)