+++ lib/knapsack_pro/rspec/test_queue_initializer.rb
@@ -14,35 +14,10 @@ module KnapsackPro
           exit 0
         end
 
-        # The test queue was created by another node while this one calculated the slow id paths via the RSpec dry-run.
-        exit 0 if test_queue_exists?
-
         initialize_test_queue(slow_id_paths)
       end
 
       private
-
-      def test_queue_exists?
-        action = KnapsackPro::Client::API::V2::Queues.connect
-        connection = KnapsackPro::Client::Connection.new(action)
-        response = connection.call
-
-        unless connection.success?
-          KnapsackPro.logger.error "Failed to initialize the test queue."
-          exit 1
-        end
-
-        if connection.errors?
-          KnapsackPro.logger.error "Failed to initialize the test queue."
-          KnapsackPro.logger.error response.inspect
-          exit 1
-        end
-
-        return false unless response.key?('url')
-
-        KnapsackPro.logger.info "Test Queue URL: #{response.fetch('url')}"
-        true
-      end
 
       def initialize_test_queue(slow_id_paths)
         all_test_files_to_run = KnapsackPro::TestSuite.new(KnapsackPro::Adapters::RSpecAdapter).all_test_files_to_run
@@ -70,6 +45,10 @@ module KnapsackPro
 
         if response.key?('url')
           KnapsackPro.logger.info "Test Queue URL: #{response.fetch('url')}"
+        end
+
+        if response.key?('recommended_node_total')
+          KnapsackPro.logger.info "Recommended node_total: #{response.fetch('recommended_node_total')}"
         end
       end
     end